Introduced logging to the BasicBuildChooser
The BuildChooser classes are responsible for choosing
slaves for builds. For debugging purposes, we would like
to log what decisions functions such as _popNextSlave
make. This will help us track down if a slave is truly
lost and help us determine how effective bug fixes are.

Introduced other logging in the AbstractLatentBuildSlave
and Builder classes.